+++ spec/dummy/client/entries/app-kitten.js
@@ -3,8 +3,8 @@ require('../stylesheets/app-kitten.scss');
 import ReactOnRails from 'react-on-rails';
 
 // Box
-import LinkBox from 'kitten/components/box/link-box'
 import DocLinkBox from 'kitten/components/box/doc-link-box'
+import LinkBox from 'kitten/components/box/link-box'
 
 // Buttons
 import ButtonImageWithTextAndBadge from 'kitten/components/buttons/button-image-with-text-and-badge'
@@ -21,10 +21,11 @@ import DropdownButton from 'kitten/components/dropdowns/dropdown-button'
 import { Checkbox } from 'kitten/components/form/checkbox'
 import FormAmountAndCurrency from 'kitten/components/form/form-amount-and-currency'
 import FormPhoneNumber from 'kitten/components/form/form-phone-number'
-import Slider from 'kitten/components/form/slider'
+import { RadioButton } from 'kitten/components/form/radio-button'
+import Search from 'kitten/components/form/search'
 import Select from 'react-select'
 import SelectWithState from 'kitten/components/form/select-with-state'
-import Search from 'kitten/components/form/search'
+import Slider from 'kitten/components/form/slider'
 import Switch from 'kitten/components/form/switch'
 
 // Grid
@@ -53,8 +54,8 @@ import TourStep from 'kitten/components/tours/tour-step'
 import { Title } from 'kitten/components/typography/title'
 
 // Simulators
-import SimulatorResult from 'kitten/components/simulators/simulator-result'
 import LoanSimulator from 'kitten/components/simulators/loan-simulator'
+import SimulatorResult from 'kitten/components/simulators/simulator-result'
 
 // Steppers
 import { Stepper } from 'kitten/components/steppers/stepper'
@@ -66,13 +67,20 @@ import { KarlImageExternalRichLink,
 import { KarlSimpleDropdown,
          KarlPlatformSwitch,
          KarlUserMenuDropdown } from 'kitten/karl/organisms/dropdown'
-import { KarlUnloggedHeaderPlatformSwitch,
+import { KarlHeader,
+         KarlHeaderWithSearchInput,
+         KarlHeaderWhenLogged,
+         KarlUnloggedHeaderPlatformSwitch,
          KarlLoggedHeaderPlatformSwitch,
          KarlLoggedHeaderUserMenu,
-         KarlPhoneDropdown } from 'kitten/karl/organisms/header'
+         KarlPhoneDropdown } from 'kitten/karl/headers/header'
 import { KarlLoanSimulator,
-         KarlLoanSimulatorWithCommission
+         KarlLoanSimulatorWithCommission,
+         KarlLoanSimulatorWithError,
        } from 'kitten/karl/organisms/loan-simulator'
+import { KarlSideGridWithRightAside,
+         KarlSideGridWithLeftAsideAndSmallSize
+       } from 'kitten/karl/grid/side-grid'
 import { KarlFormActions } from 'kitten/karl/form/form-actions'
 import { KarlGridTwelveColumns,
          KarlGridInGrid,
@@ -81,16 +89,26 @@ import { KarlGridTwelveColumns,
        } from 'kitten/karl/grid/grid'
 
 ReactOnRails.register({
+  // Box
+  DocLinkBox,
+  LinkBox,
+
+  // Buttons
   ButtonImageWithTextAndBadge,
   ButtonTooltipIcon,
+
+  // Dev
   Checkbox,
   DevGrid,
-  DocLinkBox,
+
+  // Dropdowns
   Dropdown,
   DropdownButton,
-  ExternalRichLink,
+
+  // Form
   FormAmountAndCurrency,
   FormPhoneNumber,
+  RadioButton,
   Grid,
   GridCol,
   GrabberIcon,
@@ -99,29 +117,60 @@ ReactOnRails.register({
   LoanSimulator,
   ReactTooltip,
   Slider,
+
   Search,
   Select,
   SelectWithState,
+  Slider,
+  Switch,
+
+  // Icons
+  GrabberIcon,
+
+  // Links
+  ExternalRichLink,
+
+  // Lists
   NavList,
-  SimulatorResult,
+
+  // Tooltips
+  ReactTooltip,
   Tooltip,
   Stepper,
   Switch,
   Title,
+
+  // Tours
+  HeaderTour,
   Tour,
   TourPopover,
   TourStep,
+
+  // Simulators
+  LoanSimulator,
+  SimulatorResult,
+
+  // Steppers
+  Stepper,
+
+  // Karl composed components
   KarlImageExternalRichLink,
   KarlSvgExternalRichLink,
   KarlSimpleDropdown,
   KarlPlatformSwitch,
   KarlUserMenuDropdown,
+  KarlHeader,
+  KarlHeaderWithSearchInput,
+  KarlHeaderWhenLogged,
   KarlUnloggedHeaderPlatformSwitch,
   KarlLoggedHeaderPlatformSwitch,
   KarlLoggedHeaderUserMenu,
   KarlPhoneDropdown,
   KarlLoanSimulator,
   KarlLoanSimulatorWithCommission,
+  KarlLoanSimulatorWithError,
+  KarlSideGridWithRightAside,
+  KarlSideGridWithLeftAsideAndSmallSize,
   KarlFormActions,
   KarlGridTwelveColumns,
   KarlGridInGrid,